Move data-utils to root of playwright directory
