+++ playwright-e2e/functional/consented/issue_application.spec.ts
@@ -2,7 +2,7 @@ import { test } from '../../fixtures/fixtures';
 import config from '../../config/config';
 import { ConsentedEvents } from '../../config/case-data';
 import { adminNotesTabData } from '../../resources/tab_content/consented/admin_notes_tabs';
-import { ConsentedCaseDataHelper } from '../data-utils/consented/ConsentedCaseDataHelper';
+import { ConsentedCaseDataHelper } from '../../data-utils/consented/ConsentedCaseDataHelper';
 
 test(
   'Consented - Issue Application',
