+++ playwright-e2e/functional/contested/FR_ManageHearings/manage_hearings.spec.ts
@@ -2,8 +2,8 @@ import { test } from '../../../fixtures/fixtures.ts';
 import config from '../../../config/config.ts';
 import {ContestedEvents} from "../../../config/case-data.ts";
 import {getManageHearingTableData} from "../../../resources/check_your_answer_content/manage_hearings/manageHearingAddHearingTable.ts";
-import { DateHelper } from '../../data-utils/DateHelper.ts';
-import { ContestedCaseDataHelper } from '../../data-utils/contested/ContestedCaseDataHelper.ts';
+import { DateHelper } from '../../../data-utils/DateHelper.ts';
+import { ContestedCaseDataHelper } from '../../../data-utils/contested/ContestedCaseDataHelper.ts';
 
 const typeOfHearingData = [
     "Maintenance Pending Suit (MPS)",
