+++ playwright-e2e/resources/check_your_answer_content/manage_hearings/manageHearingAddHearingTable.ts
@@ -1,4 +1,4 @@
-import { DateHelper } from "../../../functional/data-utils/DateHelper";
+import { DateHelper } from "../../../data-utils/DateHelper";
 import {Table} from "../../../pages/components/table";
 
 export function getManageHearingTableData(params?: {
