+++ playwright-e2e/resources/tab_content/contested/solicitor_create_case_tabs.ts
@@ -1,4 +1,4 @@
-import { DateHelper } from "../../../functional/data-utils/DateHelper";
+import { DateHelper } from "../../../data-utils/DateHelper";
 
 export const createCaseTabData = [
   {
